# opal_timers: ASM update