@jsquyres 2016-04-11
@bosilca @hjelmn Does @hpcraink's changes look ok?