@hpcraink 2016-04-12
@jsquyres Thanks for asking, just to give an update:
We're still working on it; George is in favor of dropping RDTSC altogether in favor of clock_gettime.
This is however RDTSC-based as well ,-) So we enter glibc, which calls the underlying clock_gettime in Linux-Kernel's vDSO (see linux/arch/x86/entry/vdso/vclock_gettime.c).
George measured on Haswell.
In the past, there have been issues with TSC; but with Invariant TSC (as all modern Intel and AMD cpus offer), the TSC should be constant regardless of power-mode, TurboBoost.
So the only thing left is the MPI process being scheduled to another core: to quote Sergey Kostrov "The smallest difference in TSC values for two different cores I've measured was 708 nano-seconds. It should be less for fast 3rd generation CPUs, like Ivy Bridge, etc"
https://software.intel.com/en-us/forums/software-tuning-performance-optimization-platform-monitoring/topic/388964
The linux kernel btw does a always return the maximum value read out, so the clock stays monotonic.
Moreover RHEL 6.7 based libc don't offer clock_gettime, yet.
RHEL-7.2 has clock_gettime:

Just checking on JUSTUS (uni-ulm's Haswell-based system):
Timing: TSC: rdtsc:48.10 rdtscp:62.11 clock_gettime (MONOTONIC):158.26 clock_gettime(REALTIME):148.86
Timing: NS: rdtsc:20.04ns rdtscp:25.88ns clock_gettime (MONOTONIC):66.41ns clock_gettime(REALTIME):61.98ns
(first line is average of 1000 measurements; second line is divided by frequency in Nanoseconds).

One solution could be to enable TSC at runtime, iff the user has bound processes to single cores!
Needless to say, again some fiddling, changing code and possibly a new PR ,-)