@jsquyres 2016-04-12
+1 for moving in the direction of `clock_gettime()`.

FWIW: RHEL 6.5 has `clock_gettime()` -- I'd be surprised if it was dropped in RHEL 6.7...?