@ibm-ompi 2016-05-20
Test passed.