@jsquyres 2016-09-26
@hpcraink Any update, perchance?  (it looks like this PR needs to be rebased, too)